==> 1.make_players_df-11to22.ipynb <==
import pandas as pd

data = pd.read_csv('data/11to22_players_list.txt', sep = "\t")

data.columns = ['c1']
data = data[data['c1'].str.contains('-')]
data

print(data['c1'].iloc[0].split('-')[0].strip())
print(data['c1'].iloc[0].split('-')[1].strip().split('(')[0].split(',')[0].lower().strip())
print(data['c1'].iloc[0].split('(')[-1][:-1].split(',')[0].strip())
print(data['c1'].iloc[0].split('(')[-1][:-1].split(',')[1].strip())
#data['c1'].iloc[0].split('(')[-1][:-1]

# for x in data['c1']:
#     #print(x)
#     print(x.split('(')[-1][:-1].split(',')[-1].strip())

def kor(x):
    k_name = x.split('-')[0].strip()
    return k_name

def eng(x):
    name = x.split('-')[1]
    return name.split('(')[0].lower().strip()

def team(x):
    return x.split('(')[-1][:-1].split(',')[-1].strip()

def get_end_dt(x):
    year = x.split('(')[-1][:-1].split(',')[0].strip()
    try:
        return f'{int(year)-1}-12-31'
    except:
        return year
    
def release_status(x):
    return 1 if '방출' in x else 0
    
data['kor_name'] = data['c1'].apply(kor)
data['eng_name'] = data['c1'].apply(eng)
data['team'] = data['c1'].apply(team)
data['release'] = data['c1'].apply(release_status)
data['end_dt'] = data['c1'].apply(get_end_dt)
data

data = data[data['kor_name']!='']
data = data.drop('c1',axis=1)
data.reset_index(inplace=True,drop=True)
data

data.to_csv('new_data/players_df.csv',index=False)
