==> 2.get_savant_data.ipynb <==
import pandas as pd
import numpy as np
import pybaseball

# Data Load

players_df = pd.read_csv('new_data/players_df.csv')
players_df

# id를 추출하기 위한 전처리

players_df.loc[players_df['kor_name']=='조조 레이예스','eng_name'] = 'jo-jo reyes'
players_df.loc[players_df['kor_name']=='왕웨이중','eng_name'] = 'wei-chung wang'

## pybaseball 라이브러리에 검색이 되지 않는 선수 확인

pybaseball.playerid_lookup('wang','wei-chung')

tmp_lst = []
for tmp in players_df['eng_name']:
    if len(pybaseball.playerid_lookup(tmp.split(' ')[-1],tmp.split(' ')[0]))!=1:
        tmp_lst.append(tmp)
        print(tmp,':',len(pybaseball.playerid_lookup(tmp.split(' ')[-1],tmp.split(' ')[0])))

# id 추출

## 마틴 이름 변경

players_df[players_df['kor_name']=='마틴']

pybaseball.playerid_lookup('martin','j. d.')

players_df.loc[players_df['kor_name']=='마틴','eng_name'] = 'j. d. martin'
players_df[players_df['kor_name']=='마틴']

def get_name_first_last(x):
    if ',' in x:
        name_first,name_last = x.split(',')[-1], x.split(',')[0]
    elif 'martin' in x:
        name_first,name_last = 'martin','j. d.'
    else:
        name_first,name_last = x.split(' ')[-1], x.split(' ')[0]
    return name_first, name_last

def get_start_dt(x):
    name_first, name_last = get_name_first_last(x)
    tmp = pybaseball.playerid_lookup(name_first,name_last)
    return str(int(tmp.iloc[-1]['mlb_played_first']))+'-01-01' if len(tmp)>0 else np.nan

def get_savant_id(x):
    name_first, name_last = get_name_first_last(x)
    tmp = pybaseball.playerid_lookup(name_first,name_last)
    return tmp.iloc[-1]['key_mlbam'] if len(tmp)>0 else np.nan

def get_fangraphs_id(x):
    name_first, name_last = get_name_first_last(x)
    tmp = pybaseball.playerid_lookup(name_first,name_last)
    return tmp.iloc[-1]['key_fangraphs'] if len(tmp)>0 else np.nan

def get_player_id(x):
    name_first, name_last = get_name_first_last(x)
    tmp = pybaseball.playerid_lookup(name_first,name_last)
    return tmp.iloc[-1]['key_bbref'] if len(tmp)>0 else np.nan

players_df['start_dt'] = players_df['eng_name'].apply(get_start_dt)
players_df['savant_id'] = players_df['eng_name'].apply(get_savant_id)
players_df['fangraphs_id'] = players_df['eng_name'].apply(get_fangraphs_id)
players_df['player_id'] = players_df['eng_name'].apply(get_player_id)

players_df

# mlb 경력없는 선수 삭제
players_df = players_df[~players_df['savant_id'].isna()].reset_index(drop=True)
players_df

players_df[players_df['kor_name']=='마틴']

## 추출 및 병합

df = pd.DataFrame()
for i in range(len(players_df)):
    tmp = players_df.loc[i]
    try:
        tmp_df = pybaseball.statcast_pitcher(tmp['start_dt'],tmp['end_dt'],tmp['savant_id'])
    except:
        print(tmp['kor_name'],' 에러')
    df = pd.concat([df,tmp_df])

df

# 파일로 저장

import pickle
import gzip

# save and compress.
with gzip.open('new_data/savant_df_zip.pickle', 'wb') as f:
    pickle.dump(df, f)

players_df_ver2 = players_df[players_df['kor_name']!='메릴 켈리']
players_df_ver2.to_csv('new_data/players_df_ver2.csv',index=False)
