==> 4.KBO_data.ipynb <==
import pandas as pd

# 동명이인 전처리 및 이름 통일화

df1 = pd.read_csv('new_data/kbo_data.csv')
df1.head()

df2 = pd.read_csv('new_data/players_df_ver2.csv')
df2.head()

df2.loc[df2['kor_name']=='채드 벨','kor_name'] = '채드벨'
df2.loc[df2['kor_name']=='팻 딘','kor_name'] = '팻딘'
df2.loc[df2['kor_name']=='트래비스','kor_name'] = '트레비스'
df2.loc[df2['kor_name']=='샘슨','kor_name'] = '키버스 샘슨'
df2.loc[df2['kor_name']=='수아레즈','kor_name'] = '앤드류 수아레즈'

name_lst1 = list(df1['이름'].unique())
name_lst2 = list(df2['kor_name'].unique())
name_lst1.append('아드리안 샘슨')
name_lst1.append('알버트 수아레즈')
name_dict = {}
for n1 in name_lst1:
    if n1 == '브룩스':
        name_dict['애런 브룩스'] = n1
        continue
    for n2 in name_lst2:
        if n1 in n2:
            if n1 in ['켈리','수아레즈','샘슨']:
                name_dict[n2] = n2
            else:
                name_dict[n2] = n1
            break

print(len(name_lst2),len(name_dict.keys()))
print(set(name_lst2)-set(name_dict.keys()))

df1[df1['팀'].str.contains('22N')]

# df1[df1['이름'].str.contains('페냐')]
df1[df1['팀'].str.contains('22한')]

test = []
for i,j in name_dict.items():
    if j in i:
        test.append('T')
    else:
        test.append('F')
print('count T: ',test.count('T'))
print('count F: ',test.count('F'))

df2['kor_name'] = df2['kor_name'].map(name_dict)
df2['kor_name'].head()

df2 = df2[df2['kor_name']==df2['kor_name']]
df2.head()

df2.to_csv('new_data/players_df_ver3.csv',index=False)

df2[df2['kor_name'].str.contains('켈리')]

df2[df2['kor_name'].str.contains('샘슨')]

df2[df2['kor_name'].str.contains('수아레즈')]

df1.loc[(df1['이름']=='켈리')&(df1['팀'].str.contains('L')),'이름'] = '케이시 켈리'

df1.loc[(df1['이름']=='샘슨')&(df1['팀'].str.contains('한')),'이름'] = '키버스 샘슨'
df1.loc[(df1['이름']=='샘슨')&(df1['팀'].str.contains('롯')),'이름'] = '아드리안 샘슨'

df1.loc[(df1['이름']=='수아레즈')&(df1['팀'].str.contains('L')),'이름'] = '앤드류 수아레즈'
df1.loc[(df1['이름']=='수아레즈')&(df1['팀'].str.contains('삼')),'이름'] = '알버트 수아레즈'

df1[df1['이름'].str.contains('켈리|샘슨|수아레즈')]

change_name = {}
for key,value in zip(df2['kor_name'],df2['eng_name']):
    change_name[key] = value
df1['이름'].map(change_name)

# kbo_df 전처리

kbo_df = df1.copy()
kbo_df.head()

## 컬럼명 변경

kbo_df.columns

new_col = ['Name', '팀', 'G', 'CG', 'ShO', 'GS', 'W', 'L', 'SV', 'HLD', 'IP', 'R', 'ER', 'TBF', 'H', '2B', '3B', 'HR', 'BB', 'IBB', 'HBP', 'SO', 'BK', 'WP', 'ERA', 'FIP', 'WHIP', 'ERA+', 'FIP+', 'WAR', 'WPA']

kbo_df.columns = new_col
kbo_df.head()

kbo_df['ENG_Name'] = kbo_df['Name'].map(change_name)
kbo_df = kbo_df[kbo_df['ENG_Name']==kbo_df['ENG_Name']]
kbo_df.head()

kbo_df['Season'] = kbo_df['팀'].map(lambda x:'20'+ x[:2])
kbo_df['Team'] = kbo_df['팀'].map(lambda x:x[2])

kbo_df = kbo_df[['Name','ENG_Name','Season', 'Team', 'G', 'CG', 'ShO', 'GS', 'W', 'L', 'SV', 'HLD', 'IP', 'R', 'ER',
       'TBF', 'H', '2B', '3B', 'HR', 'BB', 'IBB', 'HBP', 'SO', 'BK', 'WP',
       'ERA', 'FIP', 'WHIP', 'ERA+', 'FIP+', 'WAR', 'WPA']]
kbo_df.head()

kbo_df = kbo_df.drop(index=34).reset_index(drop=True)

kbo_df.to_csv('new_data/kbo_df.csv',index=False)
