==> 5.final_preprocessing.ipynb <==
import pickle
import gzip
import pandas as pd

with gzip.open('new_data/savant_df_zip.pickle','rb') as f:
    svt_df = pickle.load(f)
fan_df = pd.read_csv('new_data/fan_df.csv')
kbo_df = pd.read_csv('new_data/kbo_df.csv')


print(svt_df.shape)
svt_df.head()

print(fan_df.shape)
fan_df.head()

print(kbo_df.shape)
kbo_df.head()

def change_name(x):
    if ',' in x:
        x = x.split(',')[1].strip()+ ' ' +x.split(',')[0].strip()
    for i,j in zip(['á','é','ñ','ú','í'],['a','e','n','u','i']):
        if i in x:
            x = x.replace(i,j)
    if x == 'Sugar Ray Marimon':
        x = 'Sugar Marimon'
    if x == 'Rick Van Den Hurk':
        x = 'Rick VandenHurk'
    if x == 'Jake Brigham':
        x = 'Jacob Brigham'
    return x

svt_df['player_name'] = svt_df['player_name'].apply(change_name)

svt_name = sorted(svt_df['player_name'].unique())
fan_name = sorted(fan_df['Name'].unique())
for s,f in zip(svt_name,fan_name):
    if s!=f:
        print(s,':',f)

sorted(svt_df['player_name'].unique()) == sorted(fan_df['Name'].unique())
set(svt_df['player_name'].unique()) - set(kbo_df['ENG_Name'].unique())

svt_df = svt_df[~svt_df['player_name'].str.contains('Felix Pena|Matt Dermody')]
fan_df = fan_df[~fan_df['Name'].str.contains('Felix Pena|Matt Dermody')]
print(svt_df.shape)
print(fan_df.shape)

with gzip.open('new_data/svt_df.pickle', 'wb') as f:
    pickle.dump(svt_df, f)
fan_df.to_csv('new_data/fan_df.csv',index=False)
